refactor(lifecycle): split lifecycle.rs into lifecycle/ module directory
Split 4,889-line lifecycle.rs into per-concern submodules so each submodule
owns its own #[cfg(test)] mod tests and compiles independently:

  lifecycle/mod.rs         — LifecycleManager struct + constants + builders + run_loop/tick + LifecycleHandle + free helpers
  lifecycle/tick.rs        — poll_one + sweep_merged_worktrees + tick-level tests
  lifecycle/scm_poll.rs    — poll_scm + check_ci_failed + check_review_backlog + SCM tests
  lifecycle/transition.rs  — terminate + transition + notify_orchestrator + transition tests
  lifecycle/stuck.rs       — should_mark_stuck + check_stuck + check_merge_conflicts + stuck/conflict tests

Public surface of ao_core::lifecycle::* is unchanged. Test count: 847.
Per-module filtering now works: cargo t -p ao-core lifecycle::tick::tests
